mod: 로직 분리 작업 완료
