+++ index.html
@@ -150,7 +150,10 @@
         </div>
       </div>
     </div>
+    <script type="module" src="./src/js/logic/getLotto.js"></script>
+
     <script type="module" src="./src/js/constants/index.js"></script>
+
     <script type="module" src="./src/js/validation/index.js"></script>
 
     <script type="module" src="./src/js/LottoSheet.js"></script>
