+++ src/js/constants/index.js
@@ -1,2 +1,3 @@
 export const PRICE_UNIT = 1000;
 export const LOTTO_MAX_NUMBER = 45;
+export const LOTTO_DIGITS = 6;
